+++ src/main/java/team01/yaksutor/admin/mapper/AdMemberMapper.java
@@ -1,0 +1,4 @@
+package team01.yaksutor.admin.mapper;
+
+public interface AdMemberMapper {
+}
