--- src/main/java/team01/yaksutor/admin/medicine/dto/Ingredient.java
@@ -1,13 +1,0 @@
-package team01.yaksutor.admin.medicine.dto;
-
-import lombok.Data;
-
-@Data
-public class Ingredient {
-    private String ingredientCode;
-    private String mediCode;
-    private String regMId;
-    private String ingredient;
-    private String content;
-    private String regDate;
-}
