--- src/main/java/team01/yaksutor/admin/medicine/mapper/AdminMapper.java
@@ -1,8 +1,0 @@
-package team01.yaksutor.admin.medicine.mapper;
-
-public interface AdminMapper {
-
-    public int medicineInsert();
-
-    public String autoIncrementMediCode();
-}
