+++ src/main/java/team01/yaksutor/admin/service/AdMemberService.java
@@ -1,0 +1,4 @@
+package team01.yaksutor.admin.service;
+
+public class AdMemberService {
+}
