+++ src/main/java/team01/yaksutor/common/mapper/MemberMapper.java
@@ -1,0 +1,4 @@
+package team01.yaksutor.common.mapper;
+
+public interface MemberMapper {
+}
