+++ src/main/java/team01/yaksutor/common/service/MemberService.java
@@ -1,0 +1,4 @@
+package team01.yaksutor.common.service;
+
+public class MemberService {
+}
