+++ src/main/java/team01/yaksutor/dto/Efficacy.java
@@ -1,0 +1,12 @@
+package team01.yaksutor.dto;
+
+import lombok.Data;
+
+@Data
+public class Efficacy {
+    private String efficacyCode;
+    private String mediCode;
+    private String regMId;
+    private String efficacy;
+    private String regDate;
+}
