+++ src/main/java/team01/yaksutor/dto/Ingredient.java
@@ -1,0 +1,13 @@
+package team01.yaksutor.dto;
+
+import lombok.Data;
+
+@Data
+public class Ingredient {
+    private String ingredientCode;
+    private String mediCode;
+    private String regMId;
+    private String ingredient;
+    private String content;
+    private String regDate;
+}
