--- src/main/java/team01/yaksutor/pharmacy/member/service/MemberService.java
@@ -1,4 +1,0 @@
-package team01.yaksutor.pharmacy.member.service;
-
-public class MemberService {
-}
