+++ src/main/java/team01/yaksutor/pharmacy/service/PhMemberService.java
@@ -1,0 +1,4 @@
+package team01.yaksutor.pharmacy.service;
+
+public class PhMemberService {
+}
