+++ src/main/java/team01/yaksutor/supplier/service/SuService.java
@@ -1,0 +1,4 @@
+package team01.yaksutor.supplier.service;
+
+public class SuService {
+}
