+++ src/main/resources/mapper/admin/AdMemberMapper.xml
@@ -2,11 +2,6 @@
 <!DOCTYPE mapper
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="team01.yaksutor.admin.mapper.AdminMapper">
-
-    <select id="autoIncrementMediCode" resultType="String">
-
-    </select>
-
+<mapper namespace="team01.yaksutor.admin.mapper.AdMemberMapper">
 
 </mapper>
