--- src/main/resources/mapper/member/MemberMapper.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper
-  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-  "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="ks51team01.yaksutor.member.mapper.MemberMapper">
-
-</mapper>
