+++ src/main/resources/templates/admin/fragments/aside.html
@@ -26,7 +26,7 @@
                     </a>
                     <div class="submenu">
                         <a th:href="@{/admin/member/memberInsert}">회원 추가</a>
-                        <a th:href="@{/admin/member/memberSearchList}">회원 검색</a>
+                        <a th:href="@{#}">회원 검색</a>
                         <a th:href="@{#}">약국 추가</a>
                         <a th:href="@{#}">약국 검색</a>
                     </div>
